project/requirements.txt
streamlit
langchain
langchain-openai
langchain-community
langchain-chroma
chromadb
pypdf
python-dotenv